Grid: cell colors relative to selected area
